// frontend/src/api/generated/models/BucketDto.ts
export interface BucketDto {
    /**
     * The ID of the bucket.
     * @type {number}
     * @memberof BucketDto
     */
    id: number;
    /**
     * The name of the bucket.
     * @type {string}
     * @memberof BucketDto
     */
    name: string;
    /**
     * The URL to the RAG server.
     * @type {string}
     * @memberof BucketDto
     */
    endpoint: string;
    /**
     * The index name
     * @type {string}
     * @memberof BucketDto
     */
    indexName: string;
    /**
     * The optional headers.
     * @type {string}
     * @memberof BucketDto
     */
    headers: string;
    /**
     * Indicates whether the bucket is the user/default bucket.
     * @type {boolean}
     * @memberof BucketDto
     */
    isDefault: boolean;
    /**
     * The quota per user. Only relevant if the bucket is a user bucket/default bucket.
     * @type {number}
     * @memberof BucketDto
     */
    perUserQuota: number;
    /**
     * The allowed file types. Only relevant if the bucket is a user bucket/default bucket.
     * @type {Array<string>}
     * @memberof BucketDto
     */
    allowedFileNameExtensions?: Array<string>;
    /**
     * The bucket type.
     * @type {string}
     * @memberof BucketDto
     */
    type?: BucketDtoTypeEnum;
    /**
     * 
     * @type {BucketDtoFileSizeLimits}
     * @memberof BucketDto
     */
    fileSizeLimits: BucketDtoFileSizeLimits;
}


/**
 * @export
 */
export const BucketDtoTypeEnum = {
    General: 'general',
    User: 'user',
    Conversation: 'conversation'
} as const;
export type BucketDtoTypeEnum = typeof BucketDtoTypeEnum[keyof typeof BucketDtoTypeEnum];